% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvmapAutoConfig.R
\name{gvmapAutoConfig}
\alias{gvmapAutoConfig}
\title{gvmapAutoConfig}
\usage{
gvmapAutoConfig(heatmap_info, legend_info, color_info)
}
\arguments{
\item{heatmap_info}{list, heatmap data information}

\item{legend_info}{data frame, legend data information}

\item{color_info}{list, color information}
}
\value{
a config file to run gvmap
}
\description{
Generating gvmap config file automaticly
}
\examples{
heatmap_info <- list("heatmap_1", "heatmap_2", "heatmap_3")
legend_info <- data.frame(legend_group = c("lg1", "lg1", "lg2"),
                          name = c("Age", "Gender", "NRAS"),
                          col_theme = c("binary_col", "binary_col", "tag_col"),
                          column_num = c(2, 3, 4))
config_info <- gvmapAutoConfig(heatmap_info = heatmap_info,
                               legend_info = legend_info)


}
