% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvmapConfig.R
\name{gvmapConfig}
\alias{gvmapConfig}
\title{Read config file of gvmap}
\usage{
gvmapConfig(config_file)
}
\arguments{
\item{config_file}{character or list, the config file you input}
}
\description{
This function is used to check basic parameter in config file.
If a basic paramter in map_config is missing, the program will stop. eg. map_config, heatmap_num, legend_num
If a basic paramter in color_config is missing, it will use default color theme.
}
\examples{
config_file <- system.file('extdata', 'config.sig.yaml', package = 'gvmap')
gvmap_config <- gvmapConfig(config_file)

}
