% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvmapModAttr.R
\name{gvmapModAttr}
\alias{gvmapModAttr}
\title{gvmap modify attribute to config file}
\usage{
gvmapModAttr(attr_name, value, target, config_file)
}
\arguments{
\item{attr_name}{character, attribute name}

\item{value}{attribute value}

\item{target}{character, target of attribute}

\item{config_file}{list, config info generating by gvmapAutoConfig}
}
\value{
a config file to run gvmap
}
\description{
gvmap modify attribute to config file
}
\examples{
heatmap_info <- list("heatmap_1", "heatmap_2", "heatmap_3")
legend_info <- data.frame(legend_group = c("legend_1", "legend_1", "legend_2"),
                          name = c("Age", "Gender", "NRAS"),
                          col_theme = c("binary_col", "binary_col", "tag_col"),
                          column_num = c(2, 3, 4))
config_file <- gvmapAutoConfig(heatmap_info = heatmap_info,
                               legend_info = legend_info)

# modify heatmap_num
config_file <- gvmapModAttr(attr_name = "heatmap_num",
                            value = 2,
                            config_file = config_file)
config_file$map_config$heatmap_num

# modify legend_num
config_file <- gvmapModAttr(attr_name = "legend_num",
                            value = 3,
                            config_file = config_file)
config_file$map_config$legend_num

# modify raw_data of heatmap_1
config_file <- gvmapModAttr(attr_name = "raw_data",
                            value = FALSE, target = "heatmap_1",
                            config_file = config_file)

# modify kmer_col of heatmap_1
config_file <- gvmapModAttr(attr_name = "kmer_col",
                            value = 3,
                            target = "heatmap_1",
                            config_file = config_file)

# modify kmer_row of heatmap_2
config_file <- gvmapModAttr(attr_name = "kmer_row",
                            value = 4,
                            target = "heatmap_2",
                            config_file = config_file)

# modify split_row_name of heatmap_2
config_file <- gvmapModAttr(attr_name = "split_row_name",
                            value = c("M1001", "M4001"),
                            target = "heatmap_2",
                            config_file = config_file)

# modify kmer_col_color of heatmap_1
config_file <- gvmapModAttr(attr_name = "kmer_col_color",
                            value = "pool_col",
                            target = "heatmap_1",
                            config_file = config_file)

# modify kmer_row_color of heatmap_1
config_file <- gvmapModAttr(attr_name = "kmer_row_color",
                            value = "tag_col",
                            target = "heatmap_1",
                            config_file = config_file)

# modify dendrogram of heatmap_1
config_file <- gvmapModAttr(attr_name = "dendrogram",
                            value = "col",
                            target = "heatmap_1",
                            config_file = config_file)

# modify Rowv of heatmap_1
config_file <- gvmapModAttr(attr_name = "Rowv",
                            value = FALSE,
                            target = "heatmap_1",
                            config_file = config_file)

# modify Colv of heatmap_1
config_file <- gvmapModAttr(attr_name = "Colv",
                            value = FALSE,
                            target = "heatmap_1",
                            config_file = config_file)

# modify color_theme of heatmap_1
config_file <- gvmapModAttr(attr_name = "color_theme",
                            value = "gradient_col_gr_oo",
                            target = "heatmap_1",
                            config_file = config_file)

# modify distfun of heatmap_1
config_file <- gvmapModAttr(attr_name = "distfun",
                            value = dist,
                            target = "heatmap_1",
                            config_file = config_file)

# modify distfun of heatmap_1
config_file <- gvmapModAttr(attr_name = "hclustfun",
                            value = function(x) hclust(x,method = 'ward.D2'),
                            target = "heatmap_1",
                            config_file = config_file)

# modify legend_1
# NOTE: legend only have one attr_name
leg_value <- data.frame(name = c("age", "gender"),
                        color = "binary_col",
                        column = c(2,3))
config_file <- gvmapModAttr(attr_name = "legend",
                            value = leg_value,
                            target = "legend_1",
                            config_file = config_file)

# modify color
config_file <- gvmapModAttr(attr_name = "tag_col",
                            value = "#000000",
                            config_file = config_file)

config_file <- gvmapModAttr(attr_name = "tag_col_1",
                            value = "#FF00000",
                            config_file = config_file)


}
