% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvmap.R
\name{gvmap}
\alias{gvmap}
\title{gvmap package}
\usage{
gvmap(legend_data, heatmap_data, config_file, output_svg_name,
  convert_pdf = FALSE, convert_jpg = FALSE, plot_width = 1200,
  plot_height = 1600, stroke_width = 1, dend_stroke_width = 2,
  group_span = 10, sample_span = 0, heatmap_row_span = 0, frame = TRUE,
  frame_stroke_width = 2, sample_font_size = NULL,
  legend_font_size = NULL, font_family = "Arial")
}
\arguments{
\item{legend_data}{data frame. The basic legend data of the plot. It will contain all the
sample information. Attention: the row name of the legend data must be the sample name, or
this funciton will NOT run.}

\item{heatmap_data}{list, the heatmap data of the sample. The list name of heatmap data must
like this form: heatmap_1, heatmap_2 ... And each heatmap subset must be a martrix, with sample
name being the col name and gene set being the row name.}

\item{config_file}{character, the path of your config file. see \link[gvmap]{gvmap-config}}

\item{output_svg_name}{character, the output svg file name}

\item{convert_pdf}{bool, whether to convert svg file into pdf. Attention: If your svg file is
over 10M, it's better not convert it into pdf. This bug is still being repaired.}

\item{convert_jpg}{bool, whether to convert svg file into png.}

\item{plot_width}{number, the canvas width of your figure. Default: 1200}

\item{plot_height}{number, the canvas height of your figure. Default: 1600}

\item{stroke_width}{number, the stroke width of your legend figure. Default: 1}

\item{dend_stroke_width}{number, the dendrogram stroke width. Default: 2}

\item{group_span}{number, the distance between two groups, such as heatmap_1 and legend_1. Default: 10}

\item{sample_span}{number, the distance between sample. Default: 0}

\item{heatmap_row_span}{number, the distance between two specific colnames on heatmap. Default: 0}

\item{frame}{bool, if TRUE, the legend data has black outline. Otherwise there is no outline. Default: TRUE}

\item{frame_stroke_width}{number, the stroke width of the legend frame}

\item{sample_font_size}{number, the sample font size}

\item{legend_font_size}{number, the legend font size}

\item{font_family}{character, the font family of the plot. Default: "Arial"}
}
\description{
gvmap package
}
\examples{
browseVignettes("gvmap")

}
\seealso{
\link{heatmap}, \link[gplots]{heatmap.2}
}
